requirements.txt
aiohttp==3.8.1
aiormq==6.2.3
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
executing==0.8.1
frozenlist==1.3.0
h11==0.13.0
idna==3.3
Jinja2==3.0.3
mailjet-rest==1.3.4
MarkupSafe==2.1.0
multidict==6.0.2
pamqp==3.1.0
pydantic==1.9.0
python-dotenv==0.19.2
requests==2.27.1
simple-print==1.5.1
six==1.16.0
sniffio==1.2.0
starlette==0.18.0
termcolor==1.1.0
typing_extensions==4.1.1
ujson==5.1.0
urllib3==1.26.8
uvicorn==0.17.5
yarl==1.7.2